Add basic kernel testing framework
