+++ crates/kernel/.cargo/config.toml
@@ -1,0 +1,2 @@
+[build]
+target = "aarch64-unknown-none-softfloat"
