+++ crates/kernel/Cargo.toml
@@ -3,10 +3,6 @@ name = "kernel"
 version = "0.1.0"
 edition = "2021"
 
-[[bin]]
-name = "kernel"
-path = "src/main.rs"
-
 [dependencies]
 bitflags = "2.8.0"
 bytemuck = { version = "1.19.0", default-features = false }
