+++ crates/kernel/scripts/build-debug.sh
@@ -1,6 +1,5 @@
 #!/usr/bin/env bash
-cargo clean
-cargo rustc --target aarch64-unknown-none-softfloat -- \
-    -C link-arg=--script=./crates/kernel//script.ld \
-    -C relocation-model=static
-llvm-objcopy -O binary ../../target/aarch64-unknown-none-softfloat/debug/kernel kernel.bin
+
+set -ex
+
+PROFILE="dev" source $(dirname "$0")/build.sh
