+++ crates/kernel/scripts/build.sh
@@ -1,6 +1,22 @@
 #!/usr/bin/env bash
+
+set -ex
+
+EXAMPLE=${1-"main"}
+TARGET=aarch64-unknown-none-softfloat
+PROFILE=${PROFILE-"release"}
+
 cargo clean
-cargo rustc --release --target aarch64-unknown-none-softfloat -- \
+cargo rustc --profile=${PROFILE} --example=${EXAMPLE} \
+    --target=${TARGET} -- \
     -C link-arg=--script=./crates/kernel/script.ld \
     -C relocation-model=static
-llvm-objcopy -O binary ../../target/aarch64-unknown-none-softfloat/release/kernel kernel.bin
+
+if test "$PROFILE" = "dev" ; then
+    BINARY=../../target/${TARGET}/debug/examples/${EXAMPLE}
+else
+    BINARY=../../target/${TARGET}/${PROFILE}/examples/${EXAMPLE}
+fi
+
+cp ${BINARY} kernel.elf
+llvm-objcopy -O binary ${BINARY} kernel.bin
