+++ crates/kernel/scripts/debug.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 
-qemu-system-aarch64 -M raspi3b -dtb bcm2710-rpi-3-b-plus.dtb -display none -serial stdio -kernel kernel.bin \
-    -d int,mmu,guest_errors -s -S
+set -ex
+DEBUG_ARGS="-s -S" $(dirname "$0")/run.sh
