+++ crates/kernel/scripts/debug_ui.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+set -ex
+DEBUG_ARGS="-s -S" QEMU_DISPLAY="" $(dirname "$0")/run.sh
